Refactor code to account for new SuperImpose impl
The code has been refactored to remove the initialization of
the OpenGL context. In its place, the new static method
SuperimposerFactory::initOGL() is used, which is a simple static method
exposing the new SICAD::initOGL() static method.

The code has been cleand and polished to take into account the
code style of the other project to which it works with.