Delete any setter call of any robot interface
Up to this commit the application was overwriting any option provided
by the user to either or both Cartesian and Gaze controller.

With this commit the application no longer do this and, if asked to
move a robot arm, uses the any option provided by the user or the
default ones.